feat: rename .generated -> generated, adjust import paths, adjust copy
