+++ prisma-postgres/nextjs-starter-webpack/.gitignore
@@ -38,4 +38,4 @@ next-env.d.ts
 package-lock.json
 
 # generated Prisma client
-lib/.generated/
+lib/generated/
