+++ prisma-postgres/nextjs-starter-webpack/lib/prisma-enums.ts
@@ -1,1 +1,1 @@
-export * from './.generated/prisma/enums'
+export * from './generated/prisma/enums'
