+++ prisma-postgres/nextjs-starter-webpack/prisma/schema.prisma
@@ -4,7 +4,7 @@
 // See: https://www.prisma.io/docs/orm/prisma-schema/overview/generators#field-reference-1
 generator client {
   provider = "prisma-client"
-  output = "../lib/.generated/prisma"
+  output = "../lib/generated/prisma"
   previewFeatures = ["driverAdapters", "queryCompiler"]
 }
